Spring Security Token Provider
This change adds a SpringSecurityOAuth2TokenProvider that extracts an OAuth2
token from the Spring Security context.  In its current state, the code works
assuming a single request on the same thread that incoming request came in on.
In reality, this will almost never be the case, so this isn't functional in
most cases.  I want to keep the code around, even though we need to do some
extended research into how the use of ThreadLocals and reactive programming
intersect.

[#681]